+++ Sources/MapboxMaps/Style/AttributionDialogManager.swift
@@ -1,3 +1,5 @@
+import MapboxMobileEvents
+
 internal protocol AttributionDataSource: AnyObject {
     func attributions() -> [Attribution]
 }
@@ -21,10 +23,11 @@ internal class AttributionDialogManager {
 
     internal var isMetricsEnabled: Bool {
         get {
-            UserDefaults.standard.bool(forKey: Ornaments.metricsEnabledKey)
+            UserDefaults.mme_configuration().mme_isCollectionEnabled
         }
         set {
-            UserDefaults.standard.set(newValue, forKey: Ornaments.metricsEnabledKey)
+            UserDefaults.mme_configuration().mme_isCollectionEnabled = newValue
+            MMEEventsManager.shared().pauseOrResumeMetricsCollectionIfRequired()
         }
     }